+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
+- Added `Isekai High Elf` heritage (Elf Heritage).
+- Added `Isekai Wood Elf` heritage (Elf Heritage).
 - Added `Exceptional weapon` exceptional feats.
 - Added `Spell Master` character development feat.
 - Added `Sneaky Magic` character development feat.